+++ sapphire/sapphire-core/src/main/java/sapphire/kernel/server/KernelServerImpl.java
@@ -3,10 +3,13 @@ package sapphire.kernel.server;
 import static sapphire.runtime.Sapphire.*;
 
 import sapphire.app.AppEntryPoint;
+import sapphire.common.AppObject;
 import sapphire.common.AppObjectStub;
+import sapphire.compiler.GlobalStubConstants;
 import sapphire.kernel.client.KernelClient;
 import sapphire.kernel.common.GlobalKernelReferences;
 import sapphire.kernel.common.KernelOID;
+import sapphire.kernel.common.KernelObjectFactory;
 import sapphire.kernel.common.KernelObjectMigratingException;
 import sapphire.kernel.common.KernelObjectNotCreatedException;
 import sapphire.kernel.common.KernelObjectNotFoundException;
@@ -14,6 +17,7 @@ import sapphire.kernel.common.KernelRPCException;
 
 import sapphire.kernel.common.KernelRPC;
 import sapphire.oms.OMSServer;
+import sapphire.policy.SapphirePolicy;
 
 import java.io.Serializable;
 import java.net.InetSocketAddress;
@@ -120,6 +124,48 @@ public class KernelServerImpl implements KernelServer{
 	}
 
 	/** LOCAL INTERFACES **/
+    /**
+     * Create a replica of sapphire object.
+     * @author Venugopal Reddy K 00900280 on 18/02/18
+     * @param serverPolicyName server policy stub class name
+     * @param groupPolicyName group policy stub class name
+     * @param groupOid kernel Oid of group policy
+     * @param appObjectStub app object stub to be replicated
+     * @throws RemoteException
+     * @throws ClassNotFoundException
+     * @throws KernelObjectNotCreatedException
+     * @throws KernelObjectNotFoundException
+     */
+    public static void replicateSapphireObject(String serverPolicyName, String groupPolicyName, KernelOID groupOid, AppObjectStub appObjectStub) throws RemoteException, ClassNotFoundException, KernelObjectNotCreatedException, KernelObjectNotFoundException {
+
+		/* Create the Kernel Object for the Server Policy and get the Server Policy Stub */
+        String policyStubClassName = GlobalStubConstants.getPolicyPackageName() + "." + serverPolicyName;
+        SapphirePolicy.SapphireServerPolicy serverPolicyStub =  (SapphirePolicy.SapphireServerPolicy) KernelObjectFactory.create(policyStubClassName);
+
+		/* Get the Group Policy Stub */
+        policyStubClassName = GlobalStubConstants.getPolicyPackageName() + "." + groupPolicyName;
+        SapphirePolicy.SapphireGroupPolicy groupPolicyStub = (SapphirePolicy.SapphireGroupPolicy)KernelObjectFactory.createStubWithOid(policyStubClassName, groupOid, null);
+
+		/* Initialize the server policy and get reference */
+        SapphirePolicy.SapphireServerPolicy serverPolicy = initializeServerPolicy(serverPolicyStub);
+
+		/* Create the App Object and return the App Stub */
+        appObjectStub.$__initialize(true);
+
+		/* Initialize the server policy with app object */
+        serverPolicy.$__initialize(new AppObject(appObjectStub));
+
+		/* Inject group policy stub to server policy */
+        serverPolicy.onCreate(groupPolicyStub);
+
+        try {
+			/* Add the newly created server policy to group policy */
+            groupPolicyStub.addServer(serverPolicyStub);
+        } catch(Exception e) {
+            e.printStackTrace();
+			/* TODO: cleanup */
+        }
+    }
 	/** 
 	 * Create a new kernel object locally on this server.
 	 * 
@@ -238,7 +284,7 @@ public class KernelServerImpl implements KernelServer{
 
 		if (args.length < 4) {
 			System.out.println("Incorrect arguments to the kernel server");
-			/* Time Being for backward compatibility Region is optional in the configuration */
+			/* Time Being for backward compatability Region is optional in the configuration */
 			System.out.println("[host ip] [host port] [oms ip] [oms port] [region]");
 			return;
 		}
